+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/datapacks/serializers/StatSerializers.java
@@ -2,6 +2,7 @@ package com.robertx22.age_of_exile.database.data.stats.datapacks.serializers;
 
 import com.robertx22.age_of_exile.database.data.stats.datapacks.base.DatapackStat;
 import com.robertx22.age_of_exile.database.data.stats.datapacks.base.IStatSerializer;
+import com.robertx22.age_of_exile.database.data.stats.datapacks.stats.ConvertFromOneToOtherStat;
 import com.robertx22.age_of_exile.database.data.stats.datapacks.stats.MarkerStat;
 import com.robertx22.age_of_exile.database.data.stats.datapacks.stats.OneAppliesToOtherStat;
 import com.robertx22.age_of_exile.database.data.stats.datapacks.stats.spell_related.*;
@@ -16,6 +17,7 @@ public class StatSerializers {
 
     private StatSerializers() {
         map.put(OneAppliesToOtherStat.SER_ID, new OneAppliesToOtherSer());
+        map.put(ConvertFromOneToOtherStat.SER_ID, new TransferStatSer());
         map.put(MarkerStat.SER_ID, new MarkerSer());
 
         map.put(GiveSpellStat.SER_ID, new SpellStatSer<>(x -> new GiveSpellStat(x)));