+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/effects/offense/CriticalDamageEffect.java
@@ -10,7 +10,8 @@ public class CriticalDamageEffect extends BaseAnyEffect {
 
     @Override
     public int GetPriority() {
-        return Priority.afterThis(new CriticalHitEffect().GetPriority());
+        return Priority.afterThis(CriticalHitEffect.getInstance()
+            .GetPriority());
     }
 
     @Override