+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/offense/CriticalHit.java
@@ -22,7 +22,7 @@ public class CriticalHit extends Stat implements IStatEffects, ILocalStat {
 
     @Override
     public IStatEffect getEffect() {
-        return new CriticalHitEffect();
+        return CriticalHitEffect.getInstance();
     }
 
     private CriticalHit() {