+++ src/main/java/com/robertx22/age_of_exile/saveclasses/unit/ResourcesData.java
@@ -118,6 +118,10 @@ public class ResourcesData {
         return mana;
     }
 
+    public float getBlood() {
+        return blood;
+    }
+
     public float getMagicShield() {
         return magicShield;
     }